scripts/requirements.txt
requests==2.31.0
google-generativeai==0.8.3
python-dateutil==2.9.0
